GeneralisedJaccard Constructor to match convention.
Adjusted GeneralisedJaccard Constructor to take a threshold and
Similarity Measure at constructor time so as not to make a Similarity
Measure that will fail at runtime with a nullpointer exception when it
looks for a threshold or similarity that might not exist.

Added test and adjusted subclass to account for change in constructor.